requirements.txt
pandas
numpy
scikit-learn
pyyaml
joblib
fastapi
uvicorn
pydantic
mlflow